requirements.txt
Django~=5.0.3
djangorestframework~=3.15.1
django-cors-headers
numpy~=1.26.4
django-import-export~=4.0.0
django-environ
django-sslserver
daphne
gunicorn

